+++ CHANGES.md
@@ -8,10 +8,12 @@
 - Removed the following from `Cesium3DTilesSelection::Tile`:
   - `getLastSelectionState` and `setLastSelectionState`. Use `TilesetViewGroup::getTraversalState` instead.
   - `Tile::LoadedLinkedList`. Use `LoadedTileEnumerator` instead.
+  - `getDoNotUnloadSubtreeCount`, `incrementDoNotUnloadSubtreeCount`, `decrementDoNotUnloadSubtreeCount`, `incrementDoNotUnloadSubtreeCountOnParent`, and `decrementDoNotUnloadSubtreeCountOnParent`. Use `addReference`, `releaseReference`, and `getReferenceCount` instead.
 - The `RasterOverlayCollection` constructor now takes a `LoadedTileEnumerator` instead of a `Tile::LoadedLinkedList`.
 - `TileSelectionState` no longer uses or requires a frame number. This parameter has been removed from its various methods.
 - Derived `TilesetContentLoader` classes that aggregate other loaders must now implement `setOwnerOfNestedLoaders` to pass the owner through.
 - `DebugTileStateDatabase::recordAllTileStates` and `recordTileState` now must be given a `TilesetViewGroup` indicating which view group to record states for.
+- `ViewUpdateResult` now holds `IntrusivePointer`s to `Tile` instances rather than raw pointers.
 
 ##### Additions :tada:
 
@@ -53,6 +55,7 @@
 - Fixed bugs that could prevent valid metadata in Instanced 3D Model (i3dm) files from being parsed correctly.
 - Fixed a memory leak in `CesiumGltfReader`.
 - Fixed a bug in `ImplicitTilingUtilities::computeBoundingVolume` that incorrectly subdivided a `BoundingCylinderRegion` across the discontinuity line.
+- Fixed a broken link in the `ktx` vcpkg portfile that would cause this library to fail to build.
 
 ### v0.45.0 - 2025-03-03